+++ src/bids_model/checkContrast.m
@@ -12,9 +12,10 @@ function contrast = checkContrast(node, iCon)
 
   % (C) Copyright 2022 bidspm developers
 
-  bids.Model.validate_constrasts(node);
+  BidsModel.validate_constrasts(node);
 
-  if ~ismember(lower(node.Level), {'run', 'subject'}) && ~isTtest(node.Contrasts(iCon))
+  if ~ismember(lower(node.Level), {'run', 'session', 'subject'}) && ...
+          ~isTtest(node.Contrasts(iCon))
     notImplemented(mfilename(), 'Only t test implemented for Contrasts');
     contrast = [];
     return