+++ src/stats/utils/getRegressorIdx.m
@@ -1,4 +1,4 @@
-function  [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM)
+function  [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM, bidsSes)
   %
   % Gets from the SPM structure the regressors index corresponding
   % to the a condition convolved with the canonical HRF.
@@ -12,11 +12,13 @@ function  [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM)
   %   [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM)
   %
   % :param cdtName: name of the condition to look for
-  % :type cdtName: char or cellstr
+  % :type  cdtName: char or cellstr
   %
   % :param SPM: content of SPM.mat
-  % :type SPM: structure
+  % :type  SPM: structure
   %
+  % :param bidsSes: bids session label
+  % :type  bidsSes: char
   %
   % :returns:
   %
@@ -30,6 +32,10 @@ function  [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM)
   %
 
   % (C) Copyright 2022 bidspm developers
+
+  if nargin < 3
+    bidsSes = '';
+  end
 
   % in case the condition is of something like trial_type.foo
   tokens = regexp(cdtName, '\.', 'split');
@@ -45,6 +51,15 @@ function  [cdtName, regIdx, status] = getRegressorIdx(cdtName, SPM)
 
   regIdx = regexp(SPM.xX.name', pattern, 'match');
   regIdx = ~cellfun('isempty', regIdx);
+
+  if ~isempty(bidsSes)
+    sessToKeep = ismember({SPM.Sess.ses}, bidsSes);
+    columsnToKeep = cat(2, SPM.Sess(sessToKeep).col);
+    tmp = false(size(regIdx));
+    tmp(columsnToKeep) = regIdx(columsnToKeep);
+    regIdx = tmp;
+    clear tmp;
+  end
 
   status = checkRegressorFound(regIdx, cdtName);