+++ tests/tests_stats/utils/test_returnContrastImageFile.m
@@ -1,5 +1,4 @@
 function test_suite = test_returnContrastImageFile %#ok<*STOUT>
-
   % (C) Copyright 2022 bidspm developers
   try % assignment of 'localfunctions' is necessary in Matlab >= 2016
     test_functions = localfunctions(); %#ok<*NASGU>